fix: local deploy script problem with `generateRandomName`  (#53)
* fix: error on local deployment

* refactor: move generateName to avoid circular dependency
